Valid. PostController

+++ backend/antenna/src/main/java/com/ssafy/antenna/controller/PostController.java
@@ -1,11 +1,13 @@
 package com.ssafy.antenna.controller;
 
+import com.ssafy.antenna.domain.ErrorResponse;
 import com.ssafy.antenna.domain.ResultResponse;
-import com.ssafy.antenna.domain.comment.PostCommentReq;
+import com.ssafy.antenna.domain.comment.dto.PostCommentReq;
 import com.ssafy.antenna.domain.comment.dto.PostSubCommentReq;
 import com.ssafy.antenna.domain.post.dto.PostDetailRes;
 import com.ssafy.antenna.domain.post.dto.PostDetailWithCategory;
 import com.ssafy.antenna.domain.post.dto.PostUpdateReq;
+import com.ssafy.antenna.exception.ErrorCode;
 import com.ssafy.antenna.service.PostService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -53,7 +55,8 @@ public class PostController {
     }
 
     @GetMapping("/{postId}")
-    public ResultResponse<PostDetailWithCategory> getPostById(@PathVariable Long postId, Authentication authentication) {
+    public ResultResponse<?> getPostById(@PathVariable Long postId, Authentication authentication) {
+        if(postId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return ResultResponse.success(postService.getPostById(postId, Long.valueOf(authentication.getName())));
     }
 
@@ -63,10 +66,11 @@ public class PostController {
     }
 
     @DeleteMapping("/{postId}")
-    public ResultResponse<PostDetailRes> deletePost(
+    public ResultResponse<?> deletePost(
             @PathVariable Long postId,
             Authentication authentication
     ) throws IllegalAccessException {
+        if(postId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return ResultResponse.success(
                 postService.deletePost(
                         Long.valueOf(authentication.getName()),
@@ -76,11 +80,12 @@ public class PostController {
     }
 
     @PutMapping("/{postId}")
-    public ResultResponse<PostDetailRes> updatePost(
+    public ResultResponse<?> updatePost(
             @PathVariable Long postId,
             @RequestBody @Valid PostUpdateReq postUpdateReq,
             Authentication authentication
     ) throws IllegalAccessException {
+        if(postId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.updatePost(postId, postUpdateReq, authentication);
     }
 
@@ -90,6 +95,7 @@ public class PostController {
             @RequestBody @Valid PostCommentReq postCommentReq,
             Authentication authentication
     ) {
+        if(postId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.postComment(
                 postId,
                 postCommentReq,
@@ -98,17 +104,23 @@ public class PostController {
     }
 
     @GetMapping
-    public ResultResponse<List<PostDetailRes>> getPostWithArea(@RequestParam double lng, @RequestParam double lat, @RequestParam double area) {
+    public ResultResponse<List<PostDetailRes>> getPostWithArea(
+            @RequestParam double lng,
+            @RequestParam double lat,
+            @RequestParam double area
+    ) {
         return ResultResponse.success(postService.getPostWithArea(lng, lat, area));
     }
 
     @GetMapping("/users/{userId}")
     public ResultResponse<?> getPostByUserId(@PathVariable Long userId) {
+        if(userId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.getPostByUserId(userId);
     }
 
     @GetMapping("/{postId}/comments")
     public ResultResponse<?> getCommentsByPostId(@PathVariable Long postId) {
+        if(postId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.getCommentsByPostId(postId);
     }
 
@@ -118,6 +130,7 @@ public class PostController {
             @RequestBody @Valid PostCommentReq postCommentReq,
             Authentication authentication
     ) throws IllegalAccessException {
+        if(commentId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.updateComment(commentId, postCommentReq, Long.valueOf(authentication.getName()));
     }
 
@@ -126,6 +139,7 @@ public class PostController {
             @PathVariable Long commentId,
             Authentication authentication
     ) throws IllegalAccessException {
+        if(commentId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.deleteComment(commentId, Long.valueOf(authentication.getName()));
     }
 
@@ -134,6 +148,7 @@ public class PostController {
             @PathVariable Long postId,
             Authentication authentication
     ) {
+        if(postId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.postPostLike(postId, Long.valueOf(authentication.getName()));
     }
 
@@ -142,6 +157,7 @@ public class PostController {
             @PathVariable Long postId,
             Authentication authentication
     ) {
+        if(postId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.getPostLike(postId, Long.valueOf(authentication.getName()));
     }
 
@@ -150,6 +166,7 @@ public class PostController {
             @PathVariable Long postId,
             Authentication authentication
     ) {
+        if(postId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.deletePostLike(postId, Long.valueOf(authentication.getName()));
     }
 
@@ -158,6 +175,7 @@ public class PostController {
             @PathVariable Long commentId,
             Authentication authentication
     ) {
+        if(commentId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.postCommentLike(commentId, Long.valueOf(authentication.getName()));
     }
 
@@ -166,6 +184,7 @@ public class PostController {
             @PathVariable Long commentId,
             Authentication authentication
     ) {
+        if(commentId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.getCommentLike(commentId, Long.valueOf(authentication.getName()));
     }
 
@@ -174,29 +193,33 @@ public class PostController {
             @PathVariable Long commentId,
             Authentication authentication
     ) {
+        if(commentId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.deleteCommentLike(commentId, Long.valueOf(authentication.getName()));
     }
 
     @PostMapping("/comments/{commentId}")
     public ResultResponse<?> postSubComment(
             @PathVariable Long commentId,
-            @RequestBody PostSubCommentReq postSubCommentReq,
+            @RequestBody @Valid PostSubCommentReq postSubCommentReq,
             Authentication authentication
     ) {
+        if(commentId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.postSubComment(commentId, postSubCommentReq, Long.valueOf(authentication.getName()));
     }
 
     @GetMapping("/comments/{commentId}")
     public ResultResponse<?> getSubComments(@PathVariable Long commentId) {
+        if(commentId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.getSubComments(commentId);
     }
 
     @PutMapping("/comments/subcomments/{subCommentId}")
     public ResultResponse<?> updateSubComment(
             @PathVariable Long subCommentId,
-            @RequestBody PostSubCommentReq postSubCommentReq,
+            @RequestBody @Valid PostSubCommentReq postSubCommentReq,
             Authentication authentication
     ) throws IllegalAccessException {
+        if(subCommentId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.updateSubComment(subCommentId, postSubCommentReq, Long.valueOf(authentication.getName()));
     }
 
@@ -205,6 +228,7 @@ public class PostController {
             @PathVariable Long subCommentId,
             Authentication authentication
     ) throws IllegalAccessException {
+        if(subCommentId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.deleteSubComment(subCommentId, Long.valueOf(authentication.getName()));
     }
 
@@ -213,6 +237,7 @@ public class PostController {
             @PathVariable Long subCommentId,
             Authentication authentication
     ) {
+        if(subCommentId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.postSubCommentLike(subCommentId, Long.valueOf(authentication.getName()));
     }
 
@@ -221,6 +246,7 @@ public class PostController {
             @PathVariable Long subCommentId,
             Authentication authentication
     ) {
+        if(subCommentId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.getSubCommentLike(subCommentId, Long.valueOf(authentication.getName()));
     }
 
@@ -229,6 +255,7 @@ public class PostController {
             @PathVariable Long subCommentId,
             Authentication authentication
     ) {
+        if(subCommentId < 1) {return ResultResponse.error(ErrorResponse.of(ErrorCode.BAD_CONSTANT));}
         return postService.deleteSubCommentLike(subCommentId, Long.valueOf(authentication.getName()));
     }
